+++ src/App.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import logo from './logo.svg';
 import './App.css';
 import { ChakraProvider } from '@chakra-ui/react'
 import tema from "./temas/tema";
-import Home from "./Home";
-import Fatos from "./Fatos";
+import Home from "./pages/Home";
+import Fatos from "./pages/Fatos";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import FatoAleatorio from "./FatoAleatorio";
+import FatoAleatorio from "./pages/FatoAleatorio";
 
 const MainApp: React.FC = () => {
 
@@ -14,7 +13,6 @@ const MainApp: React.FC = () => {
         <Router>
             <Routes>
                 <Route path="/" element={<Home />} />
-                {/*<Route index element={<div>Selecione uma lista</div>} />*/}
                 <Route path="/fatos" element={<Fatos />} />
                 <Route path="/fato_aleatorio" element={<FatoAleatorio />} />
             </Routes>
